*** empty log message ***
git-svn-id: svn://svn.code.sf.net/p/clus/code/trunk@237 1a29536d-788d-45e4-adf6-63c72010a73e

+++ clus/ext/ilevelc/ILevelCInduce.java
@@ -50,7 +50,7 @@ public class ILevelCInduce extends DepthFirstInduce {
 				index[i] = new int[nb];
 				for (int j = 0; j < nb; j++) {
 					Integer value = (Integer)crIndex[i].get(j);
-					index[i][j] = value;
+					index[i][j] = value.intValue();
 				}				
 			}
 		}
